fix(windows): prevent infinite run. Fixes #11810
windows based workflows already call .Wait() in signal_windows.go,
calling it twice will result in not exiting at all.

Unix based workflows prevent that by releasing the process early on
and checking exit code of all running processes.

Windows workflows don't do this and without knowing in-depth how windows
processes work, I didn't find a way to achieve a similar "wait for all
processes" syscall.
